Tweak stock transfer permissions
Allow a user to manage a stock transfer for a source they don’t have access to when they have access to the destination and the transfer has shipped.

This commit also removes the dedicated source_location_ids and destination_location_ids as we’ve found it preferable to customize the CanCan rules by adding more specific rules rather than override the methods that return the stock location ids.