Utilize new partial
A link is passed into the partial as new_object_url does not work with
the order controller (it is a helper defined on the resource controller
while the order controller inherits from the base controller).  The
transfer_item_table does not utilize this partial as we desire a
different function on that page of just reporting the lack of any
transfer items.