Use an icon as missing image placeholder
This displays a nicer placeholder for representing a missing image on
products or variants. It uses a font icon instead of linking to a PNG
asset.

This has the advantage of being i18n friendly (the previous PNG was
english), and is slightly more convenient to use technically (doesn't
require sprockets resolving the image path). It also looks nicer.